model Order p1

+++ backend-clone-shopee/src/main/java/com/minh/shopee/domain/dto/request/OrderItemRequest.java
@@ -1,11 +1,13 @@
 package com.minh.shopee.domain.dto.request;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class OrderItemRequest {
-    private Long productId;
+    private long productId;
     private Long shopId;
-    private Double price;
+    private BigDecimal price;
     private Integer quantity;
 }
